+++ Fatras/include/ActsFatras/Kernel/detail/SimulationActor.hpp
@@ -252,7 +252,7 @@ struct SimulationActor {
   void interact(const Acts::MaterialSlab &slab, result_type &result) const {
     // run the continuous processes over a fraction of the material. returns
     // true on break condition (same as the underlying physics lists).
-    auto runContinuousPartial = [&, this](float fraction) {
+    auto runContinuousPartial = [&, this](double fraction) {
       Acts::MaterialSlab partialSlab = slab;
       partialSlab.scaleThickness(fraction);
       // material after passing this slab
@@ -303,17 +303,17 @@ struct SimulationActor {
     const float fracL0 =
         std::clamp(static_cast<float>(l0Dist / slabL0), 0.0f, 1.0f);
     // fraction of the material where the first point-like interaction occurs
-    const float frac = std::min(fracX0, fracL0);
+    const double frac = std::min(fracX0, fracL0);
 
     // do not run if there is zero material before the point-like interaction
-    if (0.0f < frac) {
+    if (0. < frac) {
       // simulate continuous processes before the point-like interaction
       if (runContinuousPartial(frac)) {
         return;
       }
     }
     // do not run if there is no point-like interaction
-    if (frac < 1.0f) {
+    if (frac < 1.) {
       // select which process to simulate
       const std::size_t process =
           (fracX0 < fracL0) ? result.x0Process : result.l0Process;
@@ -324,7 +324,7 @@ struct SimulationActor {
         return;
       }
       // simulate continuous processes after the point-like interaction
-      if (runContinuousPartial(1.0 - frac)) {
+      if (runContinuousPartial(1. - frac)) {
         return;
       }
